Tweak error handling and clean up loMode
INV_WRITE is a small lie;  as it is has the same effect as INV_READ | INV_WRITE.

Actually, I don't think the change to loWrite changes much,
as the lseek shouldn't fail unless the database connection is lost
or the fd returned an error,   in which case we handled the error
anyway.   Maybe there is a failure mode of lo_lseek I don't understand.

It seems the real difference is that we avoid a C call in the error
path.