Remark some FFI imports as unsafe
After reading the source of libpq a bit,  I marked a bunch of functions that
should almost certainly be marked as unsafe.

A few tougher calls that need benchmarks:

    PQresultErrorField:   provisionally "safe"
    PQfnumber         :   provisionally "safe"
    PQcmdTuples       :   provisionally "unsafe"

Also note that a number of "unsafe" functions can perform synchronous IO,
however these functions shouldn't ever do so in well-written programs.
These are functions such as PQfname,  which will log a message if given
an out-of-bounds index.   So don't do that,  or you'll stop the Haskell
world while the IO completes