Removed indirections that tracked explicit disconnections
Previously it had been safe to attempt to use a database connection
after it had been explicitly disconnected.  This now causes a memory
fault, which may manifest itself in benign or not-so-benign ways.

As this is intended to be a very low-level binding to libpq,  the
extra indirections added for this small behaviour seemed high.
Applications should usually be using a higher-level binding,  which
should handle this use-after-disconnect problem and other issues
in fewer indirections.