Add Joey Adams as author
Looking at the history today,  Joey really has been the driving force of
the changes over the last few months.  So even though the total contribution
might be low compared to myself and especially Grant (who still is the single
largest contributor to postgresql-libpq),  why not?   The more the merrier!