Mark PQisBusy as an unsafe call
Since postgresql-simple has started using non-blocking calls,  and
this binding is currently a little overly conservative about using safe
FFI calls,  I'm marking isBusy as unsafe as it is non-blocking.

getResult is a bit tricker as it can be a blocking call,  or not,
depending on when it's called.   sendQuery's blocking behavior
depends on PQsetnonblocking mode;  however the proper use of sendQuery
in non-blocking mode is a little opaque,  and I'll have to investigate
further.